Revert "feat: announce Homarr v2 beta (#6793)"
This reverts commit 32a51c1e60cb3b6f7f8b87006ccf800487df7f0d.
