--- .github/ISSUE_TEMPLATE/v2_beta_feedback.yml
@@ -1,131 +1,0 @@
-name: Homarr v2 beta feedback
-description: Report a specific bug, unexpected behavior, or improvement for the Homarr v2 beta
-title: "feedback(v2): "
-labels: ["needs triage"]
-body:
-  - type: markdown
-    attributes:
-      value: |
-        Thanks for testing Homarr v2. It will become the default when it is ready, and every detailed report helps us get there.
-
-        We cannot guess what you wanted from “this looks wrong.” Tell us exactly which feature is affected, what happened, why it is wrong for your use case, and what you expected instead. Add screenshots or a short recording whenever possible.
-
-        **Not useful:** “The board looks wrong.”
-
-        **Useful:** “The container changes from four columns to three after I move a widget at 1440 × 900 and 100% browser zoom. I expected it to keep four columns.”
-
-  - type: dropdown
-    id: feedback-type
-    attributes:
-      label: Feedback type
-      options:
-        - Bug
-        - Upgrade or migration
-        - Usability or design
-        - Performance
-        - Other
-    validations:
-      required: true
-
-  - type: input
-    id: version
-    attributes:
-      label: Homarr v2 version or image tag
-      description: Copy the version from the About page or provide the exact Docker image tag you tested.
-      placeholder: ghcr.io/homarr-labs/homarr-test:v2
-    validations:
-      required: true
-
-  - type: input
-    id: affected-area
-    attributes:
-      label: Affected page, feature, or widget
-      description: Name the exact page, widget, integration, container, sidebar, or action involved.
-      placeholder: "Board editor: moving a widget into a container"
-    validations:
-      required: true
-
-  - type: textarea
-    id: behavior
-    attributes:
-      label: What happened, why is it wrong, and what should happen instead?
-      description: Describe the actual result, why it does not work for you, and the exact result you expected.
-      placeholder: |
-        What happened:
-        Why this is wrong for my use case:
-        What I expected instead:
-    validations:
-      required: true
-
-  - type: textarea
-    id: reproduction
-    attributes:
-      label: Exact steps to reproduce
-      description: Start from the page you opened and list every click, keypress, drag, setting, and value needed to reproduce it.
-      placeholder: |
-        1. Open...
-        2. Set...
-        3. Drag or click...
-        4. Observe...
-    validations:
-      required: true
-
-  - type: textarea
-    id: visual-evidence
-    attributes:
-      label: Screenshots or screen recordings
-      description: Drag screenshots or a recording into this field. Include the full page for context and a close-up of the problem. Blur secrets and personal data. If you cannot add visuals, explain why.
-      placeholder: Drag screenshots or a recording here, or explain why none are available.
-    validations:
-      required: true
-
-  - type: textarea
-    id: environment
-    attributes:
-      label: Device and Homarr environment
-      description: Provide every detail that could affect the result.
-      placeholder: |
-        Device type and model:
-        Operating system and version:
-        Browser and version:
-        Installation method (Docker, Unraid, Kubernetes, other):
-        Database (SQLite, MySQL, PostgreSQL):
-        Reverse proxy and version, if used:
-    validations:
-      required: true
-
-  - type: textarea
-    id: board-context
-    attributes:
-      label: Board and display context
-      description: Include the layout details even when you are not sure they matter.
-      placeholder: |
-        Display resolution:
-        Browser zoom:
-        Desktop or mobile layout:
-        Board column count:
-        Approximate number of items:
-        Container or sidebar nesting:
-        Relevant board or widget settings:
-    validations:
-      required: true
-
-  - type: textarea
-    id: additional-information
-    attributes:
-      label: Logs and additional information
-      description: Add relevant browser console errors, Homarr logs, integration details, or anything else that may help. Remove secrets and personal information first.
-
-  - type: checkboxes
-    id: checks
-    attributes:
-      label: Before submitting
-      options:
-        - label: I added screenshots or a recording where possible, or explained why I could not.
-          required: true
-        - label: I described both what happened and the exact result I expected.
-          required: true
-        - label: I searched the existing issues for the same feedback.
-          required: true
-        - label: I removed passwords, API keys, tokens, and other private information.
-          required: true
