+++ apps/nextjs/src/app/[locale]/boards/_layout-creator.tsx
@@ -1,21 +1,17 @@
-import { createHash } from "node:crypto";
-
 import type { JSX, PropsWithChildren } from "react";
 import { notFound, redirect } from "next/navigation";
-import { cookies } from "next/headers";
+import { AppShellMain } from "@mantine/core";
 import { TRPCError } from "@trpc/server";
 
 import { getRscUserSettingsAsync } from "@homarr/api/user-server";
 import { auth } from "@homarr/auth/next";
 import { BoardProvider } from "@homarr/boards/context";
 import { EditModeProvider } from "@homarr/boards/edit-mode";
-import { dbEnv } from "@homarr/core/infrastructure/db/env";
 import { createLogger } from "@homarr/core/infrastructure/logs";
-import { v2BetaAnnouncementCookieKey } from "@homarr/definitions";
 
-import { MainHeaderContent } from "~/components/layout/header";
+import { MainHeader } from "~/components/layout/header";
 import { BoardLogoWithTitle } from "~/components/layout/logo/board-logo";
-import { V2BetaDashboardShell } from "~/components/layout/v2-beta-announcement/v2-beta-dashboard-shell";
+import { ClientShell } from "~/components/layout/shell";
 import { BoardTourGate } from "~/components/onboarding/board-tour-gate";
 import { env } from "~/env";
 import { getCurrentColorSchemeAsync } from "~/theme/color-scheme";
@@ -46,7 +42,6 @@ export const createBoardLayout = <TParams extends Params>({
   }>) => {
     const resolvedParams = await params;
     const sessionPromise = auth();
-    const cookiesPromise = cookies();
     const initialBoardPromise = getInitialBoard(resolvedParams).then(
       (board) => ({ status: "fulfilled", board }) as const,
       (error: unknown) => ({ status: "rejected", error }) as const,
@@ -63,12 +58,11 @@ export const createBoardLayout = <TParams extends Params>({
         return false;
       }
     });
-    const [session, initialBoardResult, colorScheme, shouldRunBoardTour, cookieStore] = await Promise.all([
+    const [session, initialBoardResult, colorScheme, shouldRunBoardTour] = await Promise.all([
       sessionPromise,
       initialBoardPromise,
       colorSchemePromise,
       shouldRunBoardTourPromise,
-      cookiesPromise,
     ]);
     if (initialBoardResult.status === "rejected") {
       const { error } = initialBoardResult;
@@ -89,11 +83,6 @@ export const createBoardLayout = <TParams extends Params>({
       throw error;
     }
     const initialBoard = initialBoardResult.board;
-    const viewerIdentity = session?.user.id ?? "anonymous";
-    const viewerHash = createHash("sha256").update(viewerIdentity).digest("hex").slice(0, 16);
-    const dismissalCookieName = `${v2BetaAnnouncementCookieKey}.${viewerHash}`;
-    const isAnnouncementDismissed = cookieStore.get(dismissalCookieName)?.value === "dismissed";
-    const canExportBackup = dbEnv.DRIVER === "better-sqlite3" && Boolean(session?.user.permissions.includes("admin"));
 
     return (
       <BoardProvider initialBoard={initialBoard}>
@@ -102,20 +91,14 @@ export const createBoardLayout = <TParams extends Params>({
             <BoardMantineProvider defaultColorScheme={colorScheme}>
               <CustomCss />
               <BoardTourGate enabled={shouldRunBoardTour}>
-                <V2BetaDashboardShell
-                  canExportBackup={canExportBackup}
-                  dismissalCookieName={dismissalCookieName}
-                  initiallyDismissed={isAnnouncementDismissed}
-                  header={
-                    <MainHeaderContent
-                      logo={<BoardLogoWithTitle size="md" hideTitleOnMobile />}
-                      actions={headerActions}
-                      hasNavigation={false}
-                    />
-                  }
-                >
-                  {children}
-                </V2BetaDashboardShell>
+                <ClientShell hasNavigation={false}>
+                  <MainHeader
+                    logo={<BoardLogoWithTitle size="md" hideTitleOnMobile />}
+                    actions={headerActions}
+                    hasNavigation={false}
+                  />
+                  <AppShellMain>{children}</AppShellMain>
+                </ClientShell>
               </BoardTourGate>
             </BoardMantineProvider>
           </EditModeProvider>
