+++ apps/nextjs/src/components/layout/shell.tsx
@@ -10,13 +10,11 @@ import { navigationCollapsedAtom } from "./header/burger";
 interface ClientShellProps {
   hasHeader?: boolean;
   hasNavigation?: boolean;
-  headerHeight?: number;
 }
 
 export const ClientShell = ({
   hasHeader = true,
   hasNavigation = true,
-  headerHeight = 60,
   children,
 }: PropsWithChildren<ClientShellProps>) => {
   const collapsed = useAtomValue(navigationCollapsedAtom);
@@ -25,7 +23,7 @@ export const ClientShell = ({
   return (
     <AppShell
       {...backgroundProps}
-      header={hasHeader ? { height: headerHeight } : undefined}
+      header={hasHeader ? { height: 60 } : undefined}
       navbar={
         hasNavigation
           ? {
