--- apps/nextjs/src/components/layout/v2-beta-announcement/constants.ts
@@ -1,7 +1,0 @@
-export const v2BetaBlogUrl = "https://v2.preview.homarr.dev/blog/2026/09/03/homarr-2.0";
-export const v2BetaDocsUrl = "https://v2.preview.homarr.dev/docs/getting-started/";
-export const v2BetaPreviewUrl = "https://app-v2.preview.homarr.dev";
-export const v2BetaWorkshopUrl = "https://v2.preview.homarr.dev/workshop";
-export const v2BetaFeedbackUrl = "https://github.com/homarr-labs/homarr/issues/new?template=v2_beta_feedback.yml";
-export const v2BetaDiscussionUrl = "https://github.com/homarr-labs/homarr/issues/6600";
-export const discordInviteUrl = "https://discord.com/invite/aCsmEV5RgA";
