+++ packages/definitions/src/cookie.ts
@@ -1,3 +1,2 @@
 export const colorSchemeCookieKey = "homarr.color-scheme";
 export const localeCookieKey = "homarr.locale";
-export const v2BetaAnnouncementCookieKey = "homarr.v2-beta-announcement";
